kafka-components/kafka-producer-shelf/requirements.txt
kafka-python>=2.0.2
pandas==1.5.3
pyarrow==14.0.2
redis>=5.0